Support changing the player height to a fixed height.
This will hide part of the video if the window is smaller than the fixed height (eg: 1080px).
CSS3 doesn't support `min(1080px, 100vh)` so we can't fix that easily.

We could probably use media queries to override all the rules with `playerHeight`, but I don't care
enough right now to do that since there are other scripts which can resize to a fixed size.

Eg:

... {
	height: 1080px;
}

@media (max-height: 1080px) {
	... {
		height: 100vh;
	}
}