Fix getPlayerSize()
Something wrong with the old way we get the player width/height.

We're not getting the right movie_player element. We're probably
getting an old movie_player that's no longer in the DOM, since it has
a clientWidth/clientHeight of (0,0).

So we'll just listen to window.onresize and store the width/height globally.